Basics of projectiles, to do in comments.
TODO: update portalProj when they spawn to take the angle that the portal gun is rotated at, turn that into a vector, make them travel in that direction every frame. There should only be one in existance at a time, set it as a global variable. If it exists, update its position each frame. need to create an array that walls are added to each time one is created, when projectile hits a wall, scans this array to see what wall it hit by comparing location values, then takes the angle of that wall and creates a portal of the correct color at that location with the same rotation angle as the wall it hit